@@ -17,8 +17,8 @@
   "dependencies": {
     "animate.css": "^3.5.1",
     "autoprefixer": "^6.7.2",
-    "babel-eslint": "^7.1.1",
     "babel-core": "^6.11.4",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.10",
     "babel-plugin-transform-async-to-generator": "^6.8.0",
     "babel-plugin-transform-class-properties": "^6.11.5",
@@ -32,13 +32,6 @@
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
-    "eslint": "^3.15.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-friendly-formatter": "^2.0.7",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.9.0",
     "babel-root-slash-import": "^1.1.0",
     "babel-runtime": "^6.9.2",
     "babelify": "^7.3.0",
@@ -50,6 +43,13 @@
     "css-loader": "^0.24.0",
     "del": "^2.2.1",
     "dotenv": "^2.0.0",
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-friendly-formatter": "^2.0.7",
+    "eslint-loader": "^1.6.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.9.0",
     "express": "^4.14.1",
     "extract-text-webpack-plugin": "2.0.0-beta.3",
     "file": "^0.2.2",
@@ -65,13 +65,14 @@
     "moment": "^2.14.1",
     "pace-progress": "^1.0.2",
     "postcss-loader": "^1.2.2",
-    "react": "15.4.2",
-    "react-bootstrap": "^0.30.0",
+    "prop-types": "^15.6.0",
+    "react": "^15.6.2",
+    "react-bootstrap": "^0.31.3",
     "react-bootstrap-switch": "^3.4.5",
     "react-breadcrumbs": "^1.3.16",
     "react-codemirror": "^0.2.6",
     "react-datetime": "^2.5.0",
-    "react-dom": "^15.2.1",
+    "react-dom": "^15.6.2",
     "react-hot-loader": "^3.0.0-beta.3",
     "react-onclickoutside": "^5.3.2",
     "react-router": "^2.6.0",
