@@ -13,6 +13,7 @@
 **/
 
 import React, {Component} from 'react';
+import PropTypes from 'prop-types';
 import {
   Button,
   Form,
@@ -71,7 +72,7 @@ export class BaseField extends Component {
 }
 
 BaseField.contextTypes = {
-  Form: React.PropTypes.object
+  Form: PropTypes.object
 };
 
 
