@@ -15,17 +15,28 @@
  **/
 package com.hortonworks.registries.schemaregistry;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.hortonworks.registries.schemaregistry.state.SchemaVersionLifecycleStates;
+
 import java.io.Serializable;
 
 /**
  *
  */
+@JsonIgnoreProperties(ignoreUnknown = true)
 public final class SchemaVersionInfo implements Serializable {
+
+    private static final long serialVersionUID = -669711262227194948L;
 
     /**
      * global unique id of this schema instance
      */
     private Long id;
+
+    /**
+     * schema metadata id
+     */
+    private Long schemaMetadataId;
 
     /**
      * name of this schema instance
@@ -51,18 +62,40 @@ public final class SchemaVersionInfo implements Serializable {
      * timestamp of the schema which is given in SchemaInfo
      */
     private Long timestamp;
-    
+
+    /**
+     * current stateId of this version.
+     */
+    private Byte stateId;
 
     @SuppressWarnings("unused")
     private SchemaVersionInfo() { /* Private constructor for Jackson JSON mapping */ }
 
-    public SchemaVersionInfo(Long id, String name, Integer version, String schemaText, Long timestamp, String description) {
+    public SchemaVersionInfo(Long id,
+                             String name,
+                             Integer version,
+                             String schemaText,
+                             Long timestamp,
+                             String description) {
+        this(id, name, version, null, schemaText, timestamp, description, null);
+    }
+
+    public SchemaVersionInfo(Long id,
+                             String name,
+                             Integer version,
+                             Long schemaMetadataId,
+                             String schemaText,
+                             Long timestamp,
+                             String description,
+                             Byte stateId) {
         this.id = id;
         this.name = name;
+        this.schemaMetadataId = schemaMetadataId;
         this.description = description;
         this.version = version;
         this.schemaText = schemaText;
         this.timestamp = timestamp;
+        this.stateId = stateId == null ? SchemaVersionLifecycleStates.ENABLED.getId() : stateId;
     }
 
     public Long getId() {
@@ -89,16 +122,12 @@ public final class SchemaVersionInfo implements Serializable {
         return timestamp;
     }
 
-    @Override
-    public String toString() {
-        return "SchemaVersionInfo{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", description='" + description + '\'' +
-                ", version=" + version +
-                ", schemaText='" + schemaText + '\'' +
-                ", timestamp=" + timestamp +
-                '}';
+    public Byte getStateId() {
+        return stateId;
+    }
+
+    public Long getSchemaMetadataId() {
+        return schemaMetadataId;
     }
 
     @Override
@@ -109,22 +138,40 @@ public final class SchemaVersionInfo implements Serializable {
         SchemaVersionInfo that = (SchemaVersionInfo) o;
 
         if (id != null ? !id.equals(that.id) : that.id != null) return false;
+        if (schemaMetadataId != null ? !schemaMetadataId.equals(that.schemaMetadataId) : that.schemaMetadataId != null)
+            return false;
         if (name != null ? !name.equals(that.name) : that.name != null) return false;
         if (description != null ? !description.equals(that.description) : that.description != null) return false;
         if (version != null ? !version.equals(that.version) : that.version != null) return false;
         if (schemaText != null ? !schemaText.equals(that.schemaText) : that.schemaText != null) return false;
-        return timestamp != null ? timestamp.equals(that.timestamp) : that.timestamp == null;
-
+        if (timestamp != null ? !timestamp.equals(that.timestamp) : that.timestamp != null) return false;
+        return stateId != null ? stateId.equals(that.stateId) : that.stateId == null;
     }
 
     @Override
     public int hashCode() {
         int result = id != null ? id.hashCode() : 0;
+        result = 31 * result + (schemaMetadataId != null ? schemaMetadataId.hashCode() : 0);
         result = 31 * result + (name != null ? name.hashCode() : 0);
         result = 31 * result + (description != null ? description.hashCode() : 0);
         result = 31 * result + (version != null ? version.hashCode() : 0);
         result = 31 * result + (schemaText != null ? schemaText.hashCode() : 0);
         result = 31 * result + (timestamp != null ? timestamp.hashCode() : 0);
+        result = 31 * result + (stateId != null ? stateId.hashCode() : 0);
         return result;
+    }
+
+    @Override
+    public String toString() {
+        return "SchemaVersionInfo{" +
+                "id=" + id +
+                ", schemaMetadataId=" + schemaMetadataId +
+                ", name='" + name + '\'' +
+                ", description='" + description + '\'' +
+                ", version=" + version +
+                ", schemaText='" + schemaText + '\'' +
+                ", timestamp=" + timestamp +
+                ", stateId=" + stateId +
+                '}';
     }
 }
