@@ -1,12 +1,12 @@
 /**
  * Copyright 2016 Hortonworks.
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -38,11 +38,10 @@ import com.hortonworks.registries.schemaregistry.Resourceable;
 public interface SnapshotSerializer<I, O, S> extends Resourceable {
 
     /**
-     * Returns an instance of O which is serialized input according to the given schema
+     * @param input input payload
+     * @param schema schema
      *
-     * @param input
-     * @param schema
-     * @return
+     * @return an instance of O which is serialized input according to the given schema
      */
     O serialize(I input, S schema) throws SerDesException;
 
