@@ -15,11 +15,11 @@
  **/
 package com.hortonworks.registries.schemaregistry.avro;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Maps;
 import com.hortonworks.registries.common.catalog.CatalogResponse;
 import com.hortonworks.registries.common.test.IntegrationTest;
 import com.hortonworks.registries.schemaregistry.SchemaCompatibility;
-import com.hortonworks.registries.schemaregistry.SchemaFieldQuery;
 import com.hortonworks.registries.schemaregistry.SchemaIdVersion;
 import com.hortonworks.registries.schemaregistry.SchemaMetadata;
 import com.hortonworks.registries.schemaregistry.SchemaMetadataInfo;
@@ -29,6 +29,11 @@ import com.hortonworks.registries.schemaregistry.SchemaVersionInfo;
 import com.hortonworks.registries.schemaregistry.SchemaVersionKey;
 import com.hortonworks.registries.schemaregistry.SerDesInfo;
 import com.hortonworks.registries.schemaregistry.SerDesPair;
+import com.hortonworks.registries.schemaregistry.avro.conf.SchemaRegistryTestProfileType;
+import com.hortonworks.registries.schemaregistry.avro.helper.SchemaRegistryTestServerClientWrapper;
+import com.hortonworks.registries.schemaregistry.avro.util.AvroSchemaRegistryClientUtil;
+import com.hortonworks.registries.schemaregistry.avro.util.CustomParameterizedRunner;
+import com.hortonworks.registries.schemaregistry.avro.util.SchemaRegistryTestName;
 import com.hortonworks.registries.schemaregistry.client.SchemaRegistryClient;
 import com.hortonworks.registries.schemaregistry.errors.IncompatibleSchemaException;
 import com.hortonworks.registries.schemaregistry.errors.InvalidSchemaException;
@@ -36,11 +41,17 @@ import com.hortonworks.registries.schemaregistry.errors.SchemaNotFoundException;
 import com.hortonworks.registries.schemaregistry.serdes.avro.AvroSnapshotDeserializer;
 import com.hortonworks.registries.schemaregistry.serdes.avro.AvroSnapshotSerializer;
 import com.hortonworks.registries.schemaregistry.serdes.avro.SerDesProtocolHandlerRegistry;
+import com.hortonworks.registries.schemaregistry.state.SchemaLifecycleException;
+import com.hortonworks.registries.schemaregistry.state.SchemaVersionLifecycleStateMachineInfo;
+import com.hortonworks.registries.schemaregistry.state.SchemaVersionLifecycleStateTransition;
+import com.hortonworks.registries.schemaregistry.state.SchemaVersionLifecycleStates;
 import org.apache.avro.util.Utf8;
 import org.apache.commons.io.IOUtils;
 import org.junit.Assert;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
 
 import javax.ws.rs.BadRequestException;
 import javax.ws.rs.core.Response;
@@ -51,22 +62,59 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 import static com.hortonworks.registries.common.catalog.CatalogResponse.ResponseMessage.BAD_REQUEST_PARAM_MISSING;
 import static com.hortonworks.registries.common.catalog.CatalogResponse.ResponseMessage.UNSUPPORTED_SCHEMA_TYPE;
 import static com.hortonworks.registries.schemaregistry.serdes.avro.AvroSnapshotSerializer.SERDES_PROTOCOL_VERSION;
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 
+@RunWith(CustomParameterizedRunner.class)
 @Category(IntegrationTest.class)
-public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCientTest {
+public class AvroSchemaRegistryClientTest {
 
     private static final String INVALID_SCHEMA_PROVIDER_TYPE = "invalid-schema-provider-type";
+    private static SchemaRegistryClient SCHEMA_REGISTRY_CLIENT;
+    private static SchemaRegistryTestServerClientWrapper SCHEMA_REGISTRY_TEST_SERVER_CLIENT_WRAPPER;
+    private static Map<String, Object> SCHEMA_REGISTRY_CLIENT_CONF;
 
-    /** Class to describe schema create operation failure scenarios */
+    @Rule
+    public SchemaRegistryTestName TEST_NAME_RULE = new SchemaRegistryTestName();
+
+    @CustomParameterizedRunner.Parameters
+    public static Iterable<SchemaRegistryTestProfileType> profiles() {
+        return Arrays.asList(SchemaRegistryTestProfileType.DEFAULT, SchemaRegistryTestProfileType.SSL);
+    }
+
+    @CustomParameterizedRunner.BeforeParam
+    public static void beforeParam(SchemaRegistryTestProfileType schemaRegistryTestProfileType) throws Exception {
+        SCHEMA_REGISTRY_TEST_SERVER_CLIENT_WRAPPER = new SchemaRegistryTestServerClientWrapper(schemaRegistryTestProfileType);
+        SCHEMA_REGISTRY_TEST_SERVER_CLIENT_WRAPPER.startTestServer();
+        SCHEMA_REGISTRY_CLIENT = SCHEMA_REGISTRY_TEST_SERVER_CLIENT_WRAPPER.getClient(false);
+        SCHEMA_REGISTRY_CLIENT_CONF = SCHEMA_REGISTRY_TEST_SERVER_CLIENT_WRAPPER.exportClientConf();
+    }
+
+    @CustomParameterizedRunner.AfterParam
+    public static void afterParam() throws Exception {
+        SCHEMA_REGISTRY_TEST_SERVER_CLIENT_WRAPPER.stopTestServer();
+    }
+
+    public AvroSchemaRegistryClientTest(SchemaRegistryTestProfileType schemaRegistryTestProfileType) {
+
+    }
+
+    /**
+     * Class to describe schema create operation failure scenarios
+     */
     private static class SchemaCreateFailureScenario {
         private final String test;
         private final String name;
@@ -85,7 +133,9 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
             expectedCatalogResponse = catalogResponse;
         }
 
-        /** Return true if the schema creation failed as expected */
+        /**
+         * Return true if the schema creation failed as expected
+         */
         private void testCreate(SchemaRegistryClient client) {
             boolean failedAsExpected = false;
             SchemaMetadata schemaMetadata = new SchemaMetadata.Builder(name).type(type).schemaGroup(group).
@@ -130,57 +180,8 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
     public void testSchemaCreateFailures() throws Exception {
         // Run through all the tests related to schema create failure scenarios
         for (SchemaCreateFailureScenario scenario : createFailureScenarios) {
-            scenario.testCreate(schemaRegistryClient);
+            scenario.testCreate(SCHEMA_REGISTRY_CLIENT);
         }
-    }
-
-    @Test
-    public void testSchemaOps() throws Exception {
-        SchemaMetadata schemaMetadata = createSchemaMetadata(TEST_NAME_RULE.getMethodName(), SchemaCompatibility.BOTH);
-
-        Long id = schemaRegistryClient.registerSchemaMetadata(schemaMetadata);
-        Assert.assertNotNull(id);
-
-        // registering a new schema
-        String schemaName = schemaMetadata.getName();
-        String schema1 = getSchema("/schema-1.avsc");
-        SchemaIdVersion v1 = schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(schema1, "Initial version of the schema"));
-        Assert.assertNotNull(v1.getSchemaMetadataId());
-        Assert.assertEquals(1, v1.getVersion().intValue());
-
-        SchemaMetadataInfo schemaMetadataInfoForId = schemaRegistryClient.getSchemaMetadataInfo(v1.getSchemaMetadataId());
-        SchemaMetadataInfo schemaMetadataInfoForName = schemaRegistryClient.getSchemaMetadataInfo(schemaName);
-        Assert.assertEquals(schemaMetadataInfoForId, schemaMetadataInfoForName);
-
-        // adding a new version of the schema using uploadSchemaVersion API
-        SchemaIdVersion v2 = schemaRegistryClient.uploadSchemaVersion(schemaMetadata.getName(),
-                                                                      "second version",
-                                                                      AvroSchemaRegistryClientTest.class.getResourceAsStream("/schema-2.avsc"));
-        Assert.assertEquals(v1.getVersion() + 1, v2.getVersion().intValue());
-
-        SchemaVersionInfo schemaVersionInfo = schemaRegistryClient.getSchemaVersionInfo(new SchemaVersionKey(schemaName, v2.getVersion()));
-        SchemaVersionInfo latest = schemaRegistryClient.getLatestSchemaVersionInfo(schemaName);
-        Assert.assertEquals(latest, schemaVersionInfo);
-
-        Collection<SchemaVersionInfo> allVersions = schemaRegistryClient.getAllVersions(schemaName);
-        Assert.assertEquals(2, allVersions.size());
-
-        // receive the same version as earlier without adding a new schema entry as it exists in the same schema group.
-        SchemaIdVersion version = schemaRegistryClient.addSchemaVersion(schemaMetadata, new SchemaVersion(schema1, "already added schema"));
-        Assert.assertEquals(version, v1);
-
-        Collection<SchemaVersionKey> md5SchemaVersionKeys = schemaRegistryClient.findSchemasByFields(new SchemaFieldQuery.Builder().name("md5").build());
-        Assert.assertEquals(2, md5SchemaVersionKeys.size());
-
-        Collection<SchemaVersionKey> txidSchemaVersionKeys = schemaRegistryClient.findSchemasByFields(new SchemaFieldQuery.Builder().name("txid").build());
-        Assert.assertEquals(1, txidSchemaVersionKeys.size());
-
-        // checks we can update schema meta data.
-        SchemaMetadata currentSchemaMetadata = schemaRegistryClient.getSchemaMetadataInfo(schemaName).getSchemaMetadata();
-        SchemaMetadata schemaMetadataToUpdateTo = new SchemaMetadata.Builder(currentSchemaMetadata).validationLevel(SchemaValidationLevel.LATEST).build();
-        SchemaMetadataInfo updatedSchemaMetadata = schemaRegistryClient.updateSchemaMetadata(schemaName, schemaMetadataToUpdateTo);
-
-        Assert.assertEquals(SchemaValidationLevel.LATEST, updatedSchemaMetadata.getSchemaMetadata().getValidationLevel());
     }
 
     @Test
@@ -188,27 +189,30 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
         SchemaMetadata schemaMetadata = createSchemaMetadata(TEST_NAME_RULE.getMethodName(), SchemaCompatibility.BOTH);
         String schemaName = schemaMetadata.getName();
 
-        Long id = schemaRegistryClient.registerSchemaMetadata(schemaMetadata);
-        schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(getSchema("/schema-1.avsc"), "Initial version of the schema"));
-        schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(getSchema("/schema-2.avsc"), "Second version of the schema"));
-        schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(getSchema("/schema-3.avsc"), "Third version of the schema, removes name field"));
+        Long id = SCHEMA_REGISTRY_CLIENT.registerSchemaMetadata(schemaMetadata);
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName, new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-1.avsc"), "Initial version of the schema"));
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName, new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-2.avsc"), "Second version of the schema"));
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName, new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-3.avsc"), "Third version of the schema, removes name field"));
 
         try {
-            schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(getSchema("/schema-4.avsc"), "Forth version of the schema, adds back name field, but different type"));
+            SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName, new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-4.avsc"), "Forth version of the schema, adds back name field, but different type"));
             Assert.fail("Should throw IncompatibleSchemaException as check against all schema's would find name field is not compatible with v1 and v2");
-        } catch (IncompatibleSchemaException ise){
+        } catch (IncompatibleSchemaException ise) {
             //expected
         }
 
-        SchemaMetadata currentSchemaMetadata = schemaRegistryClient.getSchemaMetadataInfo(schemaName).getSchemaMetadata();
-        SchemaMetadata schemaMetadataToUpdateTo = new SchemaMetadata.Builder(currentSchemaMetadata).validationLevel(SchemaValidationLevel.LATEST).build();
-        SchemaMetadataInfo updatedSchemaMetadata = schemaRegistryClient.updateSchemaMetadata(schemaName, schemaMetadataToUpdateTo);
+        SchemaMetadata currentSchemaMetadata = SCHEMA_REGISTRY_CLIENT.getSchemaMetadataInfo(schemaName)
+                                                                     .getSchemaMetadata();
+        SchemaMetadata schemaMetadataToUpdateTo = new SchemaMetadata.Builder(currentSchemaMetadata).validationLevel(SchemaValidationLevel.LATEST)
+                                                                                                   .build();
+        SchemaMetadataInfo updatedSchemaMetadata = SCHEMA_REGISTRY_CLIENT.updateSchemaMetadata(schemaName, schemaMetadataToUpdateTo);
 
-        Assert.assertEquals(SchemaValidationLevel.LATEST, updatedSchemaMetadata.getSchemaMetadata().getValidationLevel());
+        Assert.assertEquals(SchemaValidationLevel.LATEST, updatedSchemaMetadata.getSchemaMetadata()
+                                                                               .getValidationLevel());
 
         try {
-            schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(getSchema("/schema-4.avsc"), "Forth version of the schema, adds back name field, but different type"));
-        } catch (IncompatibleSchemaException ise){
+            SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName, new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-4.avsc"), "Forth version of the schema, adds back name field, but different type"));
+        } catch (IncompatibleSchemaException ise) {
             Assert.fail("Should not throw IncompatibleSchemaException as check against only latest schema as such should ignore v1 and v2");
         }
     }
@@ -217,22 +221,22 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
     public void testInvalidSchema() throws Exception {
         String schema = "--- invalid schema ---";
         SchemaMetadata schemaMetadata = createSchemaMetadata(TEST_NAME_RULE.getMethodName(), SchemaCompatibility.BACKWARD);
-        schemaRegistryClient.addSchemaVersion(schemaMetadata, new SchemaVersion(schema, "Initial version of the schema"));
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, new SchemaVersion(schema, "Initial version of the schema"));
     }
 
     @Test(expected = IncompatibleSchemaException.class)
     public void testIncompatibleSchemas() throws Exception {
-        String schema = getSchema("/device.avsc");
-        String incompatSchema = getSchema("/device-incompat.avsc");
+        String schema = AvroSchemaRegistryClientUtil.getSchema("/device.avsc");
+        String incompatSchema = AvroSchemaRegistryClientUtil.getSchema("/device-incompat.avsc");
 
         SchemaMetadata schemaMetadata = createSchemaMetadata(TEST_NAME_RULE.getMethodName(), SchemaCompatibility.BACKWARD);
 
         // registering a new schema
-        schemaRegistryClient.addSchemaVersion(schemaMetadata, new SchemaVersion(schema, "Initial version of the schema"));
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, new SchemaVersion(schema, "Initial version of the schema"));
 
         // adding a new version of the schema
         SchemaVersion incompatSchemaInfo = new SchemaVersion(incompatSchema, "second version");
-        schemaRegistryClient.addSchemaVersion(schemaMetadata, incompatSchemaInfo);
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, incompatSchemaInfo);
     }
 
     private SchemaMetadata createSchemaMetadata(String schemaDesc, SchemaCompatibility compatibility) {
@@ -246,20 +250,20 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
 
     @Test
     public void testDefaultSerDes() throws Exception {
-        Object defaultSerializer = schemaRegistryClient.getDefaultSerializer(AvroSchemaProvider.TYPE);
-        Object defaultDeserializer = schemaRegistryClient.getDefaultDeserializer(AvroSchemaProvider.TYPE);
+        Object defaultSerializer = SCHEMA_REGISTRY_CLIENT.getDefaultSerializer(AvroSchemaProvider.TYPE);
+        Object defaultDeserializer = SCHEMA_REGISTRY_CLIENT.getDefaultDeserializer(AvroSchemaProvider.TYPE);
         Assert.assertEquals(AvroSnapshotDeserializer.class, defaultDeserializer.getClass());
         Assert.assertEquals(AvroSnapshotSerializer.class, defaultSerializer.getClass());
     }
 
-    @Test(expected = Exception.class)
+    @Test(expected = IllegalArgumentException.class)
     public void testInvalidTypeForDefaultSer() throws Exception {
-        schemaRegistryClient.getDefaultSerializer(INVALID_SCHEMA_PROVIDER_TYPE);
+        SCHEMA_REGISTRY_CLIENT.getDefaultSerializer(INVALID_SCHEMA_PROVIDER_TYPE);
     }
 
-    @Test(expected = Exception.class)
+    @Test(expected = IllegalArgumentException.class)
     public void testInvalidTypeForDefaultDes() throws Exception {
-        schemaRegistryClient.getDefaultDeserializer(INVALID_SCHEMA_PROVIDER_TYPE);
+        SCHEMA_REGISTRY_CLIENT.getDefaultDeserializer(INVALID_SCHEMA_PROVIDER_TYPE);
     }
 
     @Test
@@ -271,7 +275,7 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
 
     private void _testAvroSerDesGenericObj(Byte protocolId) throws IOException, InvalidSchemaException, IncompatibleSchemaException, SchemaNotFoundException {
         Map<String, Object> config = Maps.newHashMap();
-        config.put(SchemaRegistryClient.Configuration.SCHEMA_REGISTRY_URL.name(), rootUrl);
+        config.putAll(SCHEMA_REGISTRY_CLIENT_CONF);
         config.put(SERDES_PROTOCOL_VERSION, protocolId);
 
         AvroSnapshotSerializer avroSnapshotSerializer = new AvroSnapshotSerializer();
@@ -279,12 +283,12 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
         AvroSnapshotDeserializer avroSnapshotDeserializer = new AvroSnapshotDeserializer();
         avroSnapshotDeserializer.init(config);
 
-        String deviceSchema = getSchema("/device.avsc");
+        String deviceSchema = AvroSchemaRegistryClientUtil.getSchema("/device.avsc");
         SchemaMetadata schemaMetadata = createSchemaMetadata(TEST_NAME_RULE.getMethodName(), SchemaCompatibility.BOTH);
-        SchemaIdVersion v1 = schemaRegistryClient.addSchemaVersion(schemaMetadata, new SchemaVersion(deviceSchema, "Initial version of the schema"));
+        SchemaIdVersion v1 = SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, new SchemaVersion(deviceSchema, "Initial version of the schema"));
         Assert.assertNotNull(v1);
 
-        Object deviceObject = createGenericRecordForDevice();
+        Object deviceObject = AvroSchemaRegistryClientUtil.createGenericRecordForDevice();
 
         byte[] serializedData = avroSnapshotSerializer.serialize(deviceObject, schemaMetadata);
         Object deserializedObj = avroSnapshotDeserializer.deserialize(new ByteArrayInputStream(serializedData), null);
@@ -300,7 +304,7 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
 
     private void _testAvroSerDesPrimitives(Byte protocolId) {
         Map<String, Object> config = Maps.newHashMap();
-        config.put(SchemaRegistryClient.Configuration.SCHEMA_REGISTRY_URL.name(), rootUrl);
+        config.putAll(SCHEMA_REGISTRY_CLIENT_CONF);
         config.put(SERDES_PROTOCOL_VERSION, protocolId);
 
         AvroSnapshotSerializer avroSnapshotSerializer = new AvroSnapshotSerializer();
@@ -308,7 +312,7 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
         AvroSnapshotDeserializer avroSnapshotDeserializer = new AvroSnapshotDeserializer();
         avroSnapshotDeserializer.init(config);
 
-        Object[] objects = generatePrimitivePayloads();
+        Object[] objects = AvroSchemaRegistryClientUtil.generatePrimitivePayloads();
         for (Object obj : objects) {
             String name = obj != null ? obj.getClass().getName() : Void.TYPE.getName();
             SchemaMetadata schemaMetadata = createSchemaMetadata(name, SchemaCompatibility.BOTH);
@@ -331,13 +335,13 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
         String fileId = uploadFile();
         SchemaMetadata schemaMetadata = createSchemaMetadata(TEST_NAME_RULE.getMethodName(), SchemaCompatibility.BOTH);
 
-        schemaRegistryClient.addSchemaVersion(schemaMetadata, new SchemaVersion(getSchema("/device.avsc"), "Initial version of the schema"));
+        SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/device.avsc"), "Initial version of the schema"));
         SerDesPair serDesPair = createSerDesInfo(fileId);
-        Long serDesId = schemaRegistryClient.addSerDes(serDesPair);
+        Long serDesId = SCHEMA_REGISTRY_CLIENT.addSerDes(serDesPair);
         Assert.assertNotNull("Returned serDesId can not be null", serDesId);
         String schemaName = schemaMetadata.getName();
-        schemaRegistryClient.mapSchemaWithSerDes(schemaName, serDesId);
-        Collection<SerDesInfo> serializers = schemaRegistryClient.getSerDes(schemaName);
+        SCHEMA_REGISTRY_CLIENT.mapSchemaWithSerDes(schemaName, serDesId);
+        Collection<SerDesInfo> serializers = SCHEMA_REGISTRY_CLIENT.getSerDes(schemaName);
 
         Assert.assertTrue(serializers.stream()
                                      .map(x -> x.getSerDesPair())
@@ -364,7 +368,221 @@ public class AvroSchemaRegistryClientTest extends AbstractAvroSchemaRegistryCien
         }
 
         InputStream inputStream = new FileInputStream(tmpJarFile);
-        return schemaRegistryClient.uploadFile(inputStream);
+        return SCHEMA_REGISTRY_CLIENT.uploadFile(inputStream);
+    }
+
+
+    @Test
+    public void testSchemaVersionDeletion() throws Exception {
+
+        SchemaVersionKey schemaVersionKey = addAndDeleteSchemaVersion(TEST_NAME_RULE.getMethodName());
+
+        Assert.assertTrue(SCHEMA_REGISTRY_CLIENT.getAllVersions(schemaVersionKey.getSchemaName()).isEmpty());
+    }
+
+    private SchemaVersionKey addAndDeleteSchemaVersion(String schemaName) throws InvalidSchemaException, IncompatibleSchemaException, SchemaNotFoundException, IOException {
+        SchemaMetadata schemaMetadata = createSchemaMetadata(schemaName, SchemaCompatibility.BOTH);
+        SchemaIdVersion schemaIdVersion = SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, new SchemaVersion(AvroSchemaRegistryClientUtil
+                                                                                                                            .getSchema("/device.avsc"), "Initial version of the schema"));
+        SchemaVersionKey schemaVersionKey = new SchemaVersionKey(schemaMetadata.getName(), schemaIdVersion.getVersion());
+        SCHEMA_REGISTRY_CLIENT.deleteSchemaVersion(schemaVersionKey);
+
+        return schemaVersionKey;
+    }
+
+    @Test(expected = SchemaNotFoundException.class)
+    public void testDeletingNonExistingSchema() throws Exception {
+
+        SchemaVersionKey schemaVersionKey = addAndDeleteSchemaVersion(TEST_NAME_RULE.getMethodName());
+
+        // deleting again should return SchemaNotFoundException
+        SCHEMA_REGISTRY_CLIENT.deleteSchemaVersion(schemaVersionKey);
+    }
+
+    @Test
+    public void testSchemaVersionEnableState() throws Exception {
+        IntStream.range(1, 100).forEach(x -> {
+            try {
+                System.out.println("########## Running for iter: " + x);
+                doTestSchemaVersionEnableState(TEST_NAME_RULE.getMethodName() + x);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+    }
+
+    public void doTestSchemaVersionEnableState(String schemaName) throws Exception {
+        String schema = AvroSchemaRegistryClientUtil.getSchema("/device.avsc");
+        String incompatSchema = AvroSchemaRegistryClientUtil.getSchema("/device-compat.avsc");
+
+        SchemaMetadata schemaMetadata = createSchemaMetadata(schemaName, SchemaCompatibility.BACKWARD);
+
+        // registering a new schema
+        SchemaIdVersion schemaIdVersion_1 = SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata,
+                                                                                    new SchemaVersion(schema, "Initial version of the schema"));
+
+        // adding a new version of the schema
+        SchemaVersion incompatSchemaInfo = new SchemaVersion(incompatSchema, "second version");
+        SchemaIdVersion schemaIdVersion_2 = SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaMetadata, incompatSchemaInfo);
+
+        Long schemaVersionId_2 = schemaIdVersion_2.getSchemaVersionId();
+        SCHEMA_REGISTRY_CLIENT.disableSchemaVersion(schemaVersionId_2);
+        SchemaVersionInfo schemaVersionInfo = SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2);
+        System.out.println("disable :: schemaVersionInfo.getStateId() = " + schemaVersionInfo.getStateId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.DISABLED.getId(), schemaVersionInfo.getStateId());
+
+        SCHEMA_REGISTRY_CLIENT.enableSchemaVersion(schemaVersionId_2);
+        schemaVersionInfo = SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2);
+        System.out.println("enable :: schemaVersionInfo.getStateId() = " + schemaVersionInfo.getStateId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.ENABLED.getId(), schemaVersionInfo.getStateId());
+
+        SCHEMA_REGISTRY_CLIENT.disableSchemaVersion(schemaVersionId_2);
+        schemaVersionInfo = SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2);
+        System.out.println("disable :: schemaVersionInfo.getStateId() = " + schemaVersionInfo.getStateId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.DISABLED.getId(), schemaVersionInfo.getStateId());
+
+        SCHEMA_REGISTRY_CLIENT.archiveSchemaVersion(schemaVersionId_2);
+        schemaVersionInfo = SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2);
+        System.out.println("archive :: schemaVersionInfo.getStateId() = " + schemaVersionInfo.getStateId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.ARCHIVED.getId(), schemaVersionInfo.getStateId());
+    }
+
+    @Test
+    public void testSchemaVersionLifeCycleStatesWithValidationAsLatest() throws Exception {
+        doTestSchemaVersionLifeCycleStates(SchemaValidationLevel.LATEST);
+    }
+
+    @Test
+    public void testSchemaVersionLifeCycleStatesWithValidationAsAll() throws Exception {
+        doTestSchemaVersionLifeCycleStates(SchemaValidationLevel.ALL);
+    }
+
+    private void doTestSchemaVersionLifeCycleStates(SchemaValidationLevel validationLevel)
+            throws InvalidSchemaException, IncompatibleSchemaException, SchemaNotFoundException, IOException, SchemaLifecycleException {
+        SchemaMetadata schemaMetadata = new SchemaMetadata.Builder(TEST_NAME_RULE.getMethodName() + "-schema")
+                .type(AvroSchemaProvider.TYPE)
+                .schemaGroup("group")
+                .compatibility(SchemaCompatibility.BOTH)
+                .validationLevel(validationLevel)
+                .build();
+        String schemaName = schemaMetadata.getName();
+
+        Long id = SCHEMA_REGISTRY_CLIENT.registerSchemaMetadata(schemaMetadata);
+
+        SchemaIdVersion schemaIdVersion_1 =
+                SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName,
+                                                        new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-1.avsc"),
+                                                                          "Initial version of the schema"));
+        SchemaIdVersion schemaIdVersion_2 =
+                SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName,
+                                                        new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-2.avsc"),
+                                                                          "Second version of the schema"));
+
+        // disable version 2
+        SCHEMA_REGISTRY_CLIENT.disableSchemaVersion(schemaIdVersion_2.getSchemaVersionId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.DISABLED.getId(),
+                            SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2).getStateId());
+
+        // add version 3
+        SchemaIdVersion schemaIdVersion_3 =
+                SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName,
+                                                        new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-3.avsc"),
+                                                                          "Third version of the schema, removes name field"));
+        // enable version 2
+        SCHEMA_REGISTRY_CLIENT.enableSchemaVersion(schemaIdVersion_2.getSchemaVersionId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.ENABLED.getId(),
+                            SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2).getStateId());
+
+        SchemaIdVersion schemaIdVersion_4 =
+                SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName,
+                                                        new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-4.avsc"),
+                                                                          "Forth version of the schema, adds back name field, but different type",
+                                                                          SchemaVersionLifecycleStates.INITIATED.getId()));
+        // enable version 4
+        try {
+            SCHEMA_REGISTRY_CLIENT.enableSchemaVersion(schemaIdVersion_4.getSchemaVersionId());
+            Assert.fail("Enabling " + schemaIdVersion_4 + " should have failed with incompatible schema error");
+        } catch (IncompatibleSchemaException e) {
+        }
+        Assert.assertEquals(SchemaVersionLifecycleStates.INITIATED.getId(),
+                            SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_4).getStateId());
+
+        // disable version 3
+        SCHEMA_REGISTRY_CLIENT.disableSchemaVersion(schemaIdVersion_3.getSchemaVersionId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.DISABLED.getId(),
+                            SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_3).getStateId());
+
+        // enable version 3
+        SCHEMA_REGISTRY_CLIENT.enableSchemaVersion(schemaIdVersion_3.getSchemaVersionId());
+        Assert.assertEquals(SchemaVersionLifecycleStates.ENABLED.getId(),
+                            SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_3).getStateId());
+    }
+
+
+    @Test
+    public void testSchemaVersionLifeCycleStateMachineConfig() throws Exception {
+        SchemaVersionLifecycleStateMachineInfo stateMachineInfo =
+                SCHEMA_REGISTRY_CLIENT.getSchemaVersionLifecycleStateMachineInfo();
+        ObjectMapper objectMapper = new ObjectMapper();
+        String stateMachineAsStr = objectMapper.writeValueAsString(stateMachineInfo);
+        SchemaVersionLifecycleStateMachineInfo readStateMachineInfo =
+                objectMapper.readValue(stateMachineAsStr, SchemaVersionLifecycleStateMachineInfo.class);
+
+        Assert.assertEquals(readStateMachineInfo, stateMachineInfo);
+
+        // check for duplicate state/transitions
+        checkDuplicateEntries(stateMachineInfo.getStates());
+        checkDuplicateEntries(stateMachineInfo.getTransitions());
+    }
+
+    private <T> void checkDuplicateEntries(Collection<T> states) {
+        HashSet<T> statesSet = new HashSet<>();
+        for (T state : states) {
+            if (!statesSet.add(state)) {
+                Assert.fail("stateMachineInfo contains duplicate state: " + state);
+            }
+        }
+    }
+
+    @Test
+    public void testSchemaVersionStatesThroughIds() throws Exception {
+        SchemaMetadata schemaMetadata = new SchemaMetadata.Builder(TEST_NAME_RULE.getMethodName() + "-schema")
+                .type(AvroSchemaProvider.TYPE)
+                .schemaGroup("group")
+                .compatibility(SchemaCompatibility.BOTH)
+                .build();
+        String schemaName = schemaMetadata.getName();
+
+        // build nextTransitions from state machine
+        SchemaVersionLifecycleStateMachineInfo stateMachine = SCHEMA_REGISTRY_CLIENT.getSchemaVersionLifecycleStateMachineInfo();
+        Map<Byte, List<SchemaVersionLifecycleStateTransition>> nextTransitionsForStateIds = new HashMap<>();
+        for (SchemaVersionLifecycleStateTransition transition : stateMachine.getTransitions()) {
+            List<SchemaVersionLifecycleStateTransition> nextTransitions = nextTransitionsForStateIds.computeIfAbsent(transition
+                                                                                                                             .getSourceStateId(),
+                                                                                                                     aByte -> new ArrayList<>());
+            nextTransitions.add(transition);
+        }
+
+        Long id = SCHEMA_REGISTRY_CLIENT.registerSchemaMetadata(schemaMetadata);
+
+        SchemaIdVersion schemaIdVersion_1 =
+                SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName,
+                                                        new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-1.avsc"),
+                                                                          "Initial version of the schema"));
+        SchemaIdVersion schemaIdVersion_2 =
+                SCHEMA_REGISTRY_CLIENT.addSchemaVersion(schemaName,
+                                                        new SchemaVersion(AvroSchemaRegistryClientUtil.getSchema("/schema-2.avsc"),
+                                                                          "Second version of the schema"));
+
+        // disable version 2
+        SchemaVersionInfo schemaVersionInfo = SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2);
+        Byte stateId = schemaVersionInfo.getStateId();
+        List<SchemaVersionLifecycleStateTransition> nextTransitions = nextTransitionsForStateIds.get(stateId);
+
+        Byte targetStateId = nextTransitions.get(0).getTargetStateId();
+        SCHEMA_REGISTRY_CLIENT.transitionState(schemaVersionInfo.getId(), targetStateId);
+
+        Assert.assertEquals(targetStateId, SCHEMA_REGISTRY_CLIENT.getSchemaVersionInfo(schemaIdVersion_2).getStateId());
     }
 
 }
