@@ -1,12 +1,12 @@
 /**
  * Copyright 2016 Hortonworks.
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,8 +15,10 @@
  **/
 package com.hortonworks.registries.schemaregistry.avro;
 
-import com.google.common.io.Resources;
-import com.hortonworks.registries.schemaregistry.webservice.LocalSchemaRegistryServer;
+import com.hortonworks.registries.schemaregistry.avro.conf.SchemaRegistryTestConfiguration;
+import com.hortonworks.registries.schemaregistry.avro.conf.SchemaRegistryTestProfileType;
+import com.hortonworks.registries.schemaregistry.avro.util.CustomParameterizedRunner;
+import com.hortonworks.registries.schemaregistry.avro.helper.SchemaRegistryTestServerClientWrapper;
 import com.hortonworks.registries.schemaregistry.SchemaIdVersion;
 import com.hortonworks.registries.schemaregistry.SchemaMetadata;
 import com.hortonworks.registries.schemaregistry.SchemaVersion;
@@ -24,41 +26,50 @@ import com.hortonworks.registries.schemaregistry.client.SchemaRegistryClient;
 import org.apache.commons.io.IOUtils;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
-import java.io.File;
-import java.util.Collections;
-import java.util.Map;
+import java.util.Arrays;
 
 /**
  *
  */
+
+@RunWith(CustomParameterizedRunner.class)
 public class LocalRegistryServerTest {
+
+    private static SchemaRegistryTestConfiguration SCHEMA_REGISTRY_TEST_CONFIGURATION;
+
+    @CustomParameterizedRunner.Parameters
+    public static Iterable<SchemaRegistryTestProfileType> profiles() {
+        return Arrays.asList(SchemaRegistryTestProfileType.DEFAULT, SchemaRegistryTestProfileType.SSL);
+    }
+
+    @CustomParameterizedRunner.BeforeParam
+    public static void setUp(SchemaRegistryTestProfileType schemaRegistryTestProfileType) throws Exception {
+        SCHEMA_REGISTRY_TEST_CONFIGURATION = SchemaRegistryTestConfiguration.forProfileType(schemaRegistryTestProfileType);
+    }
+
+    public LocalRegistryServerTest(SchemaRegistryTestProfileType schemaRegistryTestProfileType) {
+
+    }
 
     @Test
     public void testSanity() throws Exception {
-        String configPath = new File(Resources.getResource("schema-registry-test.yaml").toURI()).getAbsolutePath();
-        LocalSchemaRegistryServer localSchemaRegistryServer = new LocalSchemaRegistryServer(configPath);
-        try {
-            localSchemaRegistryServer.start();
-            SchemaRegistryClient schemaRegistryClient = createSchemaRegistryClient(localSchemaRegistryServer.getLocalPort());
 
-            // registering schema metadata
-            SchemaMetadata schemaMetadata = new SchemaMetadata.Builder("foo").type("avro").build();
-            Long schemaId = schemaRegistryClient.registerSchemaMetadata(schemaMetadata);
-            Assert.assertNotNull(schemaId);
+        SchemaRegistryTestServerClientWrapper schemaRegistryTestServerClientWrapper = new SchemaRegistryTestServerClientWrapper(SCHEMA_REGISTRY_TEST_CONFIGURATION);
+        schemaRegistryTestServerClientWrapper.startTestServer();
+        SchemaRegistryClient schemaRegistryClient = schemaRegistryTestServerClientWrapper.getClient();
 
-            // registering a new schema
-            String schemaName = schemaMetadata.getName();
-            String schema1 = IOUtils.toString(LocalRegistryServerTest.class.getResourceAsStream("/schema-1.avsc"), "UTF-8");
-            SchemaIdVersion v1 = schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(schema1, "Initial version of the schema"));
-        } finally {
-            localSchemaRegistryServer.stop();
-        }
-    }
+        // registering schema metadata
+        SchemaMetadata schemaMetadata = new SchemaMetadata.Builder("foo").type("avro").build();
+        Long schemaId = schemaRegistryClient.registerSchemaMetadata(schemaMetadata);
+        Assert.assertNotNull(schemaId);
 
-    private SchemaRegistryClient createSchemaRegistryClient(int localPort) {
-        final String rootUrl = String.format("http://localhost:%d/api/v1", localPort);
-        final Map<String, String> SCHEMA_REGISTRY_CLIENT_CONF = Collections.singletonMap(SchemaRegistryClient.Configuration.SCHEMA_REGISTRY_URL.name(), rootUrl);
-        return new SchemaRegistryClient(SCHEMA_REGISTRY_CLIENT_CONF);
+        // registering a new schema
+        String schemaName = schemaMetadata.getName();
+        String schema1 = IOUtils.toString(LocalRegistryServerTest.class.getResourceAsStream("/schema-1.avsc"), "UTF-8");
+        SchemaIdVersion v1 = schemaRegistryClient.addSchemaVersion(schemaName, new SchemaVersion(schema1, "Initial version of the schema"));
+
+        schemaRegistryTestServerClientWrapper.stopTestServer();
     }
 }
