@@ -13,55 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package com.hortonworks.registries.schemaregistry.avro;
+package com.hortonworks.registries.schemaregistry.avro.util;
 
+import com.hortonworks.registries.schemaregistry.avro.AvroSchemaRegistryClientTest;
 import com.hortonworks.registries.serdes.Device;
-import com.hortonworks.registries.schemaregistry.client.SchemaRegistryClient;
-import com.hortonworks.registries.webservice.RegistryApplication;
-import com.hortonworks.registries.webservice.RegistryConfiguration;
-import io.dropwizard.testing.ResourceHelpers;
-import io.dropwizard.testing.junit.DropwizardAppRule;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.specific.SpecificData;
 import org.apache.avro.specific.SpecificRecord;
 import org.junit.Assert;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.rules.TestName;
-
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Collections;
 import java.util.Date;
-import java.util.Map;
 import java.util.Random;
 
 /**
  *
  */
-public abstract class AbstractAvroSchemaRegistryCientTest {
+public final class AvroSchemaRegistryClientUtil {
 
-    @ClassRule
-    public static final DropwizardAppRule<RegistryConfiguration> DROPWIZARD_APP_RULE
-            = new DropwizardAppRule<>(RegistryApplication.class, ResourceHelpers.resourceFilePath("schema-registry-test.yaml"));
+    private AvroSchemaRegistryClientUtil() {
 
-    protected final String rootUrl = String.format("http://localhost:%d/api/v1", DROPWIZARD_APP_RULE.getLocalPort());
-    protected final Map<String, String> SCHEMA_REGISTRY_CLIENT_CONF = Collections.singletonMap(SchemaRegistryClient.Configuration.SCHEMA_REGISTRY_URL.name(), rootUrl);
-
-    @Rule
-    public TestName TEST_NAME_RULE = new TestName();
-
-    protected SchemaRegistryClient schemaRegistryClient;
-
-    @Before
-    public void setup() throws IOException {
-        schemaRegistryClient = new SchemaRegistryClient(SCHEMA_REGISTRY_CLIENT_CONF);
     }
 
-    protected void assertAvroObjs(Object expected, Object given) {
+    public static void assertAvroObjs(Object expected, Object given) {
         if (expected instanceof byte[]) {
             Assert.assertArrayEquals((byte[]) expected, (byte[]) given);
         } else if(expected instanceof SpecificRecord) {
@@ -71,14 +47,14 @@ public abstract class AbstractAvroSchemaRegistryCientTest {
         }
     }
 
-    protected Object[] generatePrimitivePayloads() {
+    public static Object[] generatePrimitivePayloads() {
         Random random = new Random();
         byte[] bytes = new byte[4];
         random.nextBytes(bytes);
         return new Object[]{random.nextBoolean(), random.nextDouble(), random.nextLong(), random.nextInt(), "String payload:" + new Date(), null};
     }
 
-    protected Object createGenericRecordForCompatDevice() throws IOException {
+    public static Object createGenericRecordForCompatDevice() throws IOException {
         Schema schema = new Schema.Parser().parse(getSchema("/device-compat.avsc"));
 
         GenericRecord avroRecord = new GenericData.Record(schema);
@@ -92,7 +68,7 @@ public abstract class AbstractAvroSchemaRegistryCientTest {
         return avroRecord;
     }
 
-    protected Object createGenericRecordForIncompatDevice() throws IOException {
+    public static Object createGenericRecordForIncompatDevice() throws IOException {
         Schema schema = new Schema.Parser().parse(getSchema("/device-incompat.avsc"));
 
         GenericRecord avroRecord = new GenericData.Record(schema);
@@ -105,7 +81,7 @@ public abstract class AbstractAvroSchemaRegistryCientTest {
         return avroRecord;
     }
 
-    protected Object createGenericRecordForDevice() throws IOException {
+    public static Object createGenericRecordForDevice() throws IOException {
         Schema schema = new Schema.Parser().parse(getSchema("/device.avsc"));
 
         GenericRecord avroRecord = new GenericData.Record(schema);
@@ -118,13 +94,13 @@ public abstract class AbstractAvroSchemaRegistryCientTest {
         return avroRecord;
     }
 
-    protected String getSchema(String schemaFileName) throws IOException {
+    public static String getSchema(String schemaFileName) throws IOException {
         InputStream avroSchemaStream = AvroSchemaRegistryClientTest.class.getResourceAsStream(schemaFileName);
         org.apache.avro.Schema.Parser parser = new org.apache.avro.Schema.Parser();
         return parser.parse(avroSchemaStream).toString();
     }
 
-    protected Device createSpecificRecord() {
+    public static Device createSpecificRecord() {
         return Device.newBuilder().setName("device-" + System.currentTimeMillis())
                 .setTimestamp(System.currentTimeMillis())
                 .setVersion(new Random().nextInt())
