@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.hortonworks.registries.storage.tool;
 
 import org.apache.commons.cli.BasicParser;
@@ -20,25 +21,16 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
-import org.apache.commons.cli.OptionBuilder;
 import org.apache.commons.cli.Options;
-import org.apache.commons.lang3.BooleanUtils;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.Map;
-import java.util.Scanner;
 
 public class TablesInitializer {
     private static final String OPTION_SCRIPT_ROOT_PATH = "script-root";
     private static final String OPTION_CONFIG_FILE_PATH = "config";
     private static final String OPTION_MYSQL_JAR_URL_PATH = "mysql-jar-url";
-    private static final String OPTION_EXECUTE_CREATE_TABLE = "create";
-    private static final String OPTION_EXECUTE_DROP_TABLE = "drop";
-    private static final String OPTION_EXECUTE_CHECK_CONNECTION = "check-connection";
 
-    private static final String CREATE_SCRIPT_FILE_NAME = "create_tables.sql";
-    private static final String DROP_SCRIPT_FILE_NAME = "drop_tables.sql";
 
     public static void main(String[] args) throws Exception {
         Options options = new Options();
@@ -70,24 +62,56 @@ public class TablesInitializer {
         options.addOption(
                 Option.builder()
                         .hasArg(false)
-                        .longOpt(OPTION_EXECUTE_CREATE_TABLE)
-                        .desc("Execute 'create table' script")
+                        .longOpt(SchemaMigrationOption.CREATE.toString())
+                        .desc("Run sql migrations from scatch")
                         .build()
         );
 
         options.addOption(
                 Option.builder()
                         .hasArg(false)
-                        .longOpt(OPTION_EXECUTE_DROP_TABLE)
-                        .desc("Execute 'drop table' script")
+                        .longOpt(SchemaMigrationOption.DROP.toString())
+                        .desc("Drop all the tables in the target database")
                         .build()
         );
 
         options.addOption(
                 Option.builder()
                         .hasArg(false)
-                        .longOpt(OPTION_EXECUTE_CHECK_CONNECTION)
+                        .longOpt(SchemaMigrationOption.CHECK_CONNECTION.toString())
                         .desc("Check the connection for configured data source")
+                        .build()
+        );
+
+        options.addOption(
+                Option.builder()
+                        .hasArg(false)
+                        .longOpt(SchemaMigrationOption.MIGRATE.toString())
+                        .desc("Execute schema migration from last check point")
+                        .build()
+        );
+
+        options.addOption(
+                Option.builder()
+                        .hasArg(false)
+                        .longOpt(SchemaMigrationOption.INFO.toString())
+                        .desc("Show the status of the schema migration compared to the target database")
+                        .build()
+        );
+
+        options.addOption(
+                Option.builder()
+                        .hasArg(false)
+                        .longOpt(SchemaMigrationOption.VALIDATE.toString())
+                        .desc("Validate the target database changes with the migration scripts")
+                        .build()
+        );
+
+        options.addOption(
+                Option.builder()
+                        .hasArg(false)
+                        .longOpt(SchemaMigrationOption.REPAIR.toString())
+                        .desc("Repairs the DATABASE_CHANGE_LOG by removing failed migrations and correcting checksum of existing migration script")
                         .build()
         );
 
@@ -99,19 +123,21 @@ public class TablesInitializer {
             System.exit(1);
         }
 
-        // either create or drop should be specified, not both
-        boolean executeCreate = commandLine.hasOption(OPTION_EXECUTE_CREATE_TABLE);
-        boolean executeDrop = commandLine.hasOption(OPTION_EXECUTE_DROP_TABLE);
-        boolean checkConnection = commandLine.hasOption(OPTION_EXECUTE_CHECK_CONNECTION);
+        boolean isSchemaMigrationOptionSpecified = false;
+        SchemaMigrationOption schemaMigrationOptionSpecified = null;
+        for (SchemaMigrationOption schemaMigrationOption : SchemaMigrationOption.values()) {
+            if (commandLine.hasOption(schemaMigrationOption.toString())) {
+                if (isSchemaMigrationOptionSpecified) {
+                    System.out.println("Only one operation can be execute at once, please select one of 'create', ',migrate', 'validate', 'info', 'drop', 'repair', 'check-connection'.");
+                    System.exit(1);
+                }
+                isSchemaMigrationOptionSpecified = true;
+                schemaMigrationOptionSpecified = schemaMigrationOption;
+            }
+        }
 
-        boolean moreThanOneOperationIsSpecified = executeCreate == executeDrop ? executeCreate : checkConnection;
-        boolean noOperationSpecified = !(executeCreate || executeDrop || checkConnection);
-
-        if (moreThanOneOperationIsSpecified) {
-            System.out.println("Only one operation can be execute at once, please select 'create' or 'drop', or 'check-connection'.");
-            System.exit(1);
-        } else if (noOperationSpecified) {
-            System.out.println("One of 'create', 'drop', 'check-connection' operation should be specified to execute.");
+        if (!isSchemaMigrationOptionSpecified) {
+            System.out.println("One of the option 'create', ',migrate', 'validate', 'info', 'drop', 'repair', 'check-connection' must be specified to execute.");
             System.exit(1);
         }
 
@@ -141,73 +167,15 @@ public class TablesInitializer {
             throw new IllegalStateException("Shouldn't reach here");
         }
 
+        SchemaMigrationHelper schemaMigrationHelper = new SchemaMigrationHelper(FlywayFactory.get(storageProperties, scriptRootPath));
         try {
-            SQLScriptRunner sqlScriptRunner = new SQLScriptRunner(storageProperties);
-
-            try {
-                sqlScriptRunner.initializeDriver();
-            } catch (ClassNotFoundException e) {
-                System.err.println("Driver class is not found in classpath. Please ensure that driver is in classpath.");
-                System.exit(1);
-            }
-
-            if (checkConnection) {
-                if (!sqlScriptRunner.checkConnection()) {
-                    System.exit(1);
-                }
-            } else if (executeDrop) {
-                doExecuteDrop(sqlScriptRunner, storageProperties, scriptRootPath);
-            } else {
-                // executeCreate
-                doExecuteCreate(sqlScriptRunner, storageProperties, scriptRootPath);
-            }
-        } catch (IOException e) {
-            System.err.println("Error occurred while reading script file. Script root path: " + scriptRootPath);
+            schemaMigrationHelper.execute(schemaMigrationOptionSpecified);
+            System.out.println(String.format("\"%s\" option successful", schemaMigrationOptionSpecified.toString()));
+        } catch (Exception e) {
+            System.err.println(String.format("\"%s\" option failed : %s", schemaMigrationOptionSpecified.toString(), e.getMessage()));
             System.exit(1);
         }
-    }
 
-    private static void doExecuteCreate(SQLScriptRunner sqlScriptRunner, StorageProviderConfiguration storageProperties,
-                                        String scriptRootPath) throws Exception {
-        String scriptPath = scriptRootPath + File.separator + storageProperties.getDbType() +
-                File.separator + CREATE_SCRIPT_FILE_NAME;
-
-        doExecute(sqlScriptRunner, scriptPath);
-    }
-
-    private static void doExecuteDrop(SQLScriptRunner sqlScriptRunner, StorageProviderConfiguration storageProperties,
-                                      String scriptRootPath) throws Exception {
-        System.out.println("The operation will drop any existing tables.");
-        System.out.print("Are you sure you want to proceed. (y/n)?");
-        Scanner scan = new Scanner(System.in);
-        String line = scan.nextLine();
-        System.out.println();
-
-        Boolean proceed = BooleanUtils.toBooleanObject(line);
-        if (!BooleanUtils.toBoolean(proceed)) {
-            System.exit(0);
-        }
-
-        String scriptPath = scriptRootPath + File.separator + storageProperties.getDbType() +
-                File.separator + DROP_SCRIPT_FILE_NAME;
-
-        doExecute(sqlScriptRunner, scriptPath);
-    }
-
-    private static void doExecute(SQLScriptRunner sqlScriptRunner, String scriptPath) throws Exception {
-        sqlScriptRunner.runScript(scriptPath);
-    }
-
-    private static Option option(int argCount, String shortName, String longName, String description){
-        return option(argCount, shortName, longName, longName, description);
-    }
-
-    private static Option option(int argCount, String shortName, String longName, String argName, String description){
-        return OptionBuilder.hasArgs(argCount)
-                .withArgName(argName)
-                .withLongOpt(longName)
-                .withDescription(description)
-                .create(shortName);
     }
 
     private static void usage(Options options) {
