@@ -12,7 +12,7 @@
   * limitations under the License.
 **/
 
-import React, {Component, PropTypes} from 'react';
+import React, {Component} from 'react';
 import routes from './routers/routes';
 import {render} from 'react-dom';
 import {Router, browserHistory, hashHistory} from 'react-router';
